requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
flask-cors==4.0.1
flask-bcrypt==1.0.1
pymysql==1.1.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2
pillow==10.4.0
gunicorn
psycopg2-binary